requirements.txt

streamlit>=1.36.0
google-generativeai>=0.5.0
chromadb>=0.5.0
pdfplumber>=0.11.0
requests>=2.31.0
beautifulsoup4>=4.12.3
python-dotenv>=1.0.1
sentence-transformers>=2.7.0
